+++ common/src/main/java/com/iamkaf/liteminer/LiteminerClient.java
@@ -1,18 +1,22 @@
 package com.iamkaf.liteminer;
 
+import com.iamkaf.amber.api.event.v1.events.common.client.ClientTickEvents;
+import com.iamkaf.amber.api.event.v1.events.common.client.HudEvents;
+import com.iamkaf.amber.api.event.v1.events.common.client.InputEvents;
+import com.iamkaf.amber.api.event.v1.events.common.client.RenderEvents;
+import com.iamkaf.amber.api.keymapping.KeybindHelper;
 import com.iamkaf.liteminer.config.LiteminerClientConfig;
+import com.iamkaf.liteminer.networking.C2SVeinmineKeybindChange;
 import com.iamkaf.liteminer.networking.LiteminerNetwork;
+import com.iamkaf.liteminer.rendering.BlockHighlightRenderer;
 import com.iamkaf.liteminer.rendering.HUD;
 import com.iamkaf.liteminer.shapes.Cycler;
 import com.iamkaf.liteminer.shapes.Walker;
 import com.mojang.blaze3d.platform.InputConstants;
-import dev.architectury.event.events.client.ClientGuiEvent;
-import dev.architectury.event.events.client.ClientRawInputEvent;
-import dev.architectury.event.events.client.ClientTickEvent;
-import dev.architectury.registry.client.keymappings.KeyMappingRegistry;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.phys.HitResult;
 import net.neoforged.neoforge.common.ModConfigSpec;
 import org.apache.commons.lang3.tuple.Pair;
@@ -22,11 +26,10 @@ import java.util.HashSet;
 
 public class LiteminerClient {
     public static final int PACKET_DELAY = 125;
-    public static final KeyMapping KEY_MAPPING = new KeyMapping("key.liteminer.veinmine",
-            InputConstants.Type.KEYSYM,
-            GLFW.GLFW_KEY_GRAVE_ACCENT,
-            "key.categories.liteminer"
-    );
+    public static final KeyMapping.Category KEY_CATEGORY =
+            KeyMapping.Category.register(ResourceLocation.fromNamespaceAndPath("liteminer", "liteminer"));
+    public static final KeyMapping KEY_MAPPING =
+            new KeyMapping("key.liteminer.veinmine", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_GRAVE_ACCENT, KEY_CATEGORY);
     public static final LiteminerClientConfig CONFIG;
     public static final ModConfigSpec CONFIG_SPEC;
     public static HashSet<BlockPos> selectedBlocks = HashSet.newHashSet(0);
@@ -35,20 +38,20 @@ public class LiteminerClient {
     private static long lastChange = System.currentTimeMillis();
 
     static {
-        Pair<LiteminerClientConfig, ModConfigSpec> pair =
-                new ModConfigSpec.Builder().configure(LiteminerClientConfig::new);
+        Pair<LiteminerClientConfig, ModConfigSpec> pair = new ModConfigSpec.Builder().configure(LiteminerClientConfig::new);
         CONFIG = pair.getLeft();
         CONFIG_SPEC = pair.getRight();
     }
 
     public static void init() {
-        KeyMappingRegistry.register(KEY_MAPPING);
-        ClientTickEvent.CLIENT_POST.register(LiteminerClient::onPostTick);
-        ClientGuiEvent.RENDER_HUD.register(HUD::onRenderHUD);
-        ClientRawInputEvent.MOUSE_SCROLLED.register(HUD::onMouseScroll);
+        KeybindHelper.register(KEY_MAPPING);
+        ClientTickEvents.END_CLIENT_TICK.register(LiteminerClient::onPostTick);
+        HudEvents.RENDER_HUD.register(HUD::onRenderHUD);
+        InputEvents.MOUSE_SCROLL_PRE.register(HUD::onMouseScroll);
+        RenderEvents.BLOCK_OUTLINE_RENDER.register(BlockHighlightRenderer::renderLiteminerHighlight);
     }
 
-    public static void onPostTick(Minecraft minecraft) {
+    public static void onPostTick() {
         if ((System.currentTimeMillis() - getLastChange()) < PACKET_DELAY) {
             return;
         }
@@ -61,17 +64,14 @@ public class LiteminerClient {
                     return;
                 }
 
-                new LiteminerNetwork.Messages.C2SVeinmineKeybindChange(newState,
-                        shapes.getCurrentIndex()
-                ).sendToServer();
+                LiteminerNetwork.sendToServer(new C2SVeinmineKeybindChange(newState, shapes.getCurrentIndex()));
                 currentState = newState;
             }
             case TOGGLE -> {
                 if (KEY_MAPPING.consumeClick()) {
                     var newState = !isVeinMining();
-                    new LiteminerNetwork.Messages.C2SVeinmineKeybindChange(newState,
-                            shapes.getCurrentIndex()
-                    ).sendToServer();
+
+                    LiteminerNetwork.sendToServer(new C2SVeinmineKeybindChange(newState, shapes.getCurrentIndex()));
                     currentState = newState;
                 }
             }