+++ common/src/main/java/com/iamkaf/liteminer/tags/TagHelper.java
@@ -1,9 +1,6 @@
 package com.iamkaf.liteminer.tags;
 
-import com.iamkaf.liteminer.Liteminer;
-import dev.architectury.event.events.common.LifecycleEvent;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -11,11 +8,6 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class TagHelper {
-    private static boolean initialized = false;
-
-    public static void init() {
-        LifecycleEvent.SERVER_LEVEL_LOAD.register(TagHelper::report);
-    }
 
     public static boolean isExcludedBlock(BlockState block) {
         return isOr(block, LiteminerTags.Blocks.EXCLUDED_BLOCKS, LiteminerTags.Compat.EXCLUDED_BLOCKS);
@@ -60,26 +52,5 @@ public class TagHelper {
             }
         }
         return false;
-    }
-
-    public static void report(ServerLevel serverLevel) {
-        if (initialized) return;
-        initialized = true;
-        Liteminer.LOGGER.info("Loading Liteminer Tags...");
-        BuiltInRegistries.BLOCK.get(LiteminerTags.Blocks.BLOCK_WHITELIST).ifPresent(holders -> {
-            Liteminer.LOGGER.info("Found " + holders.size() + " blocks [" + LiteminerTags.Blocks.BLOCK_WHITELIST.location() + "]");
-        });
-
-        BuiltInRegistries.BLOCK.get(LiteminerTags.Blocks.EXCLUDED_BLOCKS).ifPresent(holders -> {
-            Liteminer.LOGGER.info("Found " + holders.size() + " blocks [" + LiteminerTags.Blocks.EXCLUDED_BLOCKS.location() + "]");
-        });
-
-        BuiltInRegistries.ITEM.get(LiteminerTags.Items.EXCLUDED_TOOLS).ifPresent(holders -> {
-            Liteminer.LOGGER.info("Found " + holders.size() + " items [" + LiteminerTags.Items.EXCLUDED_TOOLS.location() + "]");
-        });
-
-        BuiltInRegistries.ITEM.get(LiteminerTags.Items.INCLUDED_TOOLS).ifPresent(holders -> {
-            Liteminer.LOGGER.info("Found " + holders.size() + " items [" + LiteminerTags.Items.INCLUDED_TOOLS.location() + "]");
-        });
     }
 }