+++ fabric/src/main/java/com/iamkaf/liteminer/fabric/client/LiteminerFabricClient.java
@@ -2,11 +2,9 @@ package com.iamkaf.liteminer.fabric.client;
 
 import com.iamkaf.liteminer.Liteminer;
 import com.iamkaf.liteminer.LiteminerClient;
-import com.iamkaf.liteminer.rendering.BlockHighlightRenderer;
 import fuzs.forgeconfigapiport.fabric.api.v5.ConfigRegistry;
 import fuzs.forgeconfigapiport.fabric.api.v5.client.ConfigScreenFactoryRegistry;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents;
 import net.neoforged.fml.config.ModConfig;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 
@@ -14,9 +12,6 @@ public final class LiteminerFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         LiteminerClient.init();
-
-        WorldRenderEvents.BLOCK_OUTLINE.register((worldRenderContext, blockOutlineContext) -> BlockHighlightRenderer.renderLiteminerHighlight(
-                worldRenderContext.matrixStack()));
 
         ConfigScreenFactoryRegistry.INSTANCE.register(Liteminer.MOD_ID, ConfigurationScreen::new);
         ConfigRegistry.INSTANCE.register(Liteminer.MOD_ID, ModConfig.Type.CLIENT, LiteminerClient.CONFIG_SPEC);