+++ neoforge/src/main/java/com/iamkaf/liteminer/neoforge/LiteminerNeoForgeClient.java
@@ -2,14 +2,10 @@ package com.iamkaf.liteminer.neoforge;
 
 import com.iamkaf.liteminer.Liteminer;
 import com.iamkaf.liteminer.LiteminerClient;
-import com.iamkaf.liteminer.rendering.BlockHighlightRenderer;
 import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModContainer;
-import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.config.ModConfig;
-import net.neoforged.neoforge.client.event.RenderHighlightEvent;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 
@@ -20,13 +16,4 @@ public class LiteminerNeoForgeClient {
         container.registerConfig(ModConfig.Type.CLIENT, LiteminerClient.CONFIG_SPEC);
         LiteminerClient.init();
     }
-
-    @EventBusSubscriber(modid = Liteminer.MOD_ID, value = Dist.CLIENT)
-    public static class Events {
-        @SubscribeEvent
-        public static void renderBlockHighlights(RenderHighlightEvent.Block event) {
-            event.setCanceled(!BlockHighlightRenderer.renderLiteminerHighlight(event.getPoseStack()));
-        }
-    }
 }
-